requirements.txt
streamlit
numpy 
pandas
yfinance
base64
matplotlib.pyplot 
seaborn